Separate player into libplayer and player
