+++ CMakeLists.txt
@@ -66,6 +66,7 @@ find_package(Sanitizers)
 
 option(BUILD_LIBAUDIO "build audio output library" ON)
 option(BUILD_LIBEMU "build sound emulation library" ON)
+option(BUILD_LIBPLAYER "build player library" ON)
 option(BUILD_TESTS "build test programs" ON)
 option(BUILD_PLAYER "build player application" ON)
 
@@ -78,7 +79,7 @@ if(BUILD_LIBEMU)
 	add_subdirectory(emu)
 	add_sanitizers(vgm-emu)
 endif()
-if(BUILD_PLAYER)
+if(BUILD_LIBPLAYER)
 	add_subdirectory(player)
 	add_sanitizers(vgm-player)
 endif()
@@ -103,7 +104,7 @@ add_sanitizers(audemutest)
 
 find_package(ZLIB REQUIRED)
 
-add_executable(vgmtest vgmtest.c vgm/dblk_compr.c)
+add_executable(vgmtest vgmtest.c player/dblk_compr.c)
 target_include_directories(vgmtest PRIVATE ${CMAKE_SOURCE_DIR})
 target_link_libraries(vgmtest PRIVATE ZLIB::ZLIB vgm-audio vgm-emu)
 add_sanitizers(vgmtest)
@@ -111,7 +112,7 @@ add_sanitizers(vgmtest)
 endif(BUILD_TESTS)
 
 if(BUILD_PLAYER)
-add_executable(player player.cpp vgm/dblk_compr.c)
+add_executable(player player.cpp player/dblk_compr.c)
 target_include_directories(player PRIVATE ${CMAKE_SOURCE_DIR})
 target_link_libraries(player PRIVATE vgm-audio vgm-player)
 add_sanitizers(player)
